Transpile dependent inclusion validator
Given a field of type <LookupName> and a parent 'from' field,
validate against acceptable values in a list for a
name, ensure that a validation is generated for it.

Also add some AST helper functions to an AST::Presenter.
Move ASTSimplifier to AST::Simplifier so that AST-related things have
a namespace.